requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4

langchain==0.1.16
langchain-community==0.0.36

requests==2.31.0
python-dotenv==1.0.1

transformers==4.38.2
torch==2.1.2

regex==2023.12.25
